Fix issue#3254: Redeclaration error related to function declaration in module top-level
Function declarations in module top-level should be treated as a lexical declarations,
and early error should be thrown upon redeclaration. Fix by checking for function
declaration in module top-level in redeclaration code path.

15.2.1.1Static Semantics: Early Errors
It is a Syntax Error if any element of the LexicallyDeclaredNames of ModuleItemList also
occurs in the VarDeclaredNames of ModuleItemList

15.2.1.11 NOTE 2
At the top level of a Module, function declarations are treated like lexical declarations
rather than like var declarations.