[CVE-2017-11802] Chakra - JIT RegexHelper::StringReplace must call the callback function with updating ImplicitCallFlags.
JIT inline string.prototype.replace for case a.replace(b,c) if 'a' and 'c' are strings and 'b' is a regex and calls helper for other cases. If c is a function RegexHelper::StringReplace will call c, this should be marked as implicit call so that JIT can bailout on implicit call check.