Merged PR 873340: Add support for a RecyclerVisitedHostHeapBlock
Add support for a RecyclerVisitedHostHeapBlock
This adds a new heap block type, and a new heap block implementation. Currently this simply derives from SmallFinalizableHeapBlockT and only has minor specialization:
SetAttributes has slightly different behavior since we can have non-finalizable objects (traced only)
HeapBlockType is set appropriately by a New/ctor implementation

Recycler macros are introduced that will allocate a chunk of memory with the appropriate bits but will not construct an object - this memory will eventually be given to the caller to put a v-table on it in order to get the visitation callbacks.

Note that true leaf objects (of which we expect to have a few in edgehtml) are not supported in this heap block type - the idea is that they will go into the existing leaf heap block