Fix split-scope special symbol access via scope-slot
If we have a split-scope with a default argument accessing a special
symbol via a scope-slot, the slot init for the special symbol needs to
happen before we emit that default argument. Change here splits the
special symbol scope slot init from the `Load*Object` methods and
performs that init when it should actually happen. Either we do it
before emitting the default arguments (in split-scope) or we do it as
part of ordinary symbol init as part of `DefineUserVars`. This greatly
simplifies our scope slot allocation and handling logic for the special
symbols. We now no longer need to manually call `EnsureScopeSlot` for
the special symbols and do not have to be as careful about ordering
their loads, inits, and stores.