Merged PR 895883: commit 5ef48746: Change the location of the caller of IRecyclerVisitedObject::Trace from UpdateAttributes to ProcessMark
This allows us to not have recursive marking for our precisely traced objects and makes it easier to reason
about what operation the callback that Trace uses (IRecyclerHeapMarkingContext::MarkObjects) performs (it just
marks every pointer that it is given.

IRecyclerHeapMarkingContext is now published, and is the new argument to IRecyclerVisitedObject::Trace.
It is implemented by a stack wrapper on MarkContext that captures the parallel and interior
template parameters from MarkContext::ProcessMark. We always use false for doSpecialMark, since all calls to Mark
that originate from ProcessMark would pass false (the only place we pass true is when scanning the stack)

IRecyclerHeapMarkingContext::MarkObjects currently does an interior mark on the assumption that the pointers provided
by the externally precisely traced object are going to be majority interior

We also changed to LeafBit designation to be more accurate, logically. Precisely traced objects are not actually leaf
since we're tracing them, and since they go into our new heap block, they will not be conservatively scanned regardless
of the leaf bit. Removing this causes us to have a couple more heap bucket template instantiations, since we now have 4
unique bit patterns to account for when masking with GetBlockTypeBitMask

Update the test to trace through a non-recyclervisited stack root, that also uses an interior pointer (but aligned) to make
sure the redirection to our heap block map for (aligned) interior pointers during non-interior marking is working as expected.