SetPropertyWithAttributes assert does not cover static constructor case
Fixes OS: 12681861

We have a condition in an assert in SetPropertyWithAttributes that checks to see if we are coming from an InitClassMember opcode. We check this by getting the constructor property on the class and checking if it has the constructor flag set.

This misses the case where we have a static method named 'constructor'. In this case we will have the class method flag set. I added this case to the assert.