Merged PR 902922: Get Object #1 (WebGLActiveInfo) onto the chakra heap for tracing and finalizing
Create macros that allow for CBase and other types of objects to use the templated ThreadAlloc's for allocation.
Use the macro for WebGLActiveInfo.
Convert AddRef/Release to RootAddRef/RootRelease calls
Virtual all non-public ref counting methods on CBase and abandon if they are called
neuter JSBind_* methods related to refcounting
Implement a Trace that simply traces the Var (which we actually know will already be marked due to how this object is reachable)
Add a basic leak detector to the new type (interim, more as a sanity check)
Modify callers creating this type of object to just new into a raw pointer on the stack.

Call Passivate (if exists) in Dispose
Ensure pointer is 16-byte aligned so that it is marked from CEO Var

Patch up the IRecyclerVisitedObject interface to not take Recycler, but instead encapsulate that inside FinalizableObject
Remove assert that prevented interior marking when in parallel/concurrent and recycler state not set (recycler state currently only used by MemGC)

Modify assert in CBase::Passivate that expects Var to be nullptr by opting out GC traced objects