Merged PR 913868: Add support for RecyclerVisitedObjects in GCStress
Add support for RecyclerVisitedObjects in GCStress
Add a new object type that is weighted similarly to TrackedObject.
The new object type will go into the new heap block type, and can be a combination of leaf, finalizable, and traced (the type of the created object is random).
We also allocate an unmanaged resource for finalizable objects to ensure that finalize is called and we release the memory (we'll eventually OOM otherwise)
Adding this showed that our new heap block type needed its own rescan logic - traced objects must go into the precisely traced mark stack instead of
being handled generically by being added to the conservatively scanned mark stack.