cleanup unit tests for arm64 with nonative
A few tests that expect jitted code and needed exclude_nonative were using require_backend or had nothing to disable the tests.

One test was for asmjs which is not currently compiled for arm or arm64.

One test was hitting stack overflow, which seems to happen at a lower threashold on arm64.

A few slower tests were timing out before completing.