Add `IsDerivedClassConstructor` and refactor `Parser::CreateSpecialSymbolDeclarations`
`IsDerivedClassConstructor` is just a convenience method for
`(IsClassConstructor && !IsBaseClassConstructor)`.

`Parser::CreateSpecialSymbolDeclarations` had the code which does the
check for special symbol references factored-out so it's a bit simpler.

Simplify the logic to construct the special symbol var declarations a
little bit. Also fix emitting a special symbol as a call target. This
fixes:
https://microsoft.visualstudio.com/OS/_workitems?id=13977724&_a=edit