Notify host Dynamic Module ready
Previously the call to notify the Host that a module was ready was made only for:
a) Errors AND
b) Parent level modules

A module loaded with a dynamic import may not be a parent and hopefully won't always be an error - this condition change means that the notification is made for any dynamic import when completed.

Also adjusts a test bug_issue_3257 which was failing to find two files - this failure was hidden by not notifying the host when a dynamic module was ready as the dynamic imports that should have failed were acting as no-ops.

Also adds a new test to ensure that dynamic imports from modules do notify the host.