Fix #3816 Class methods cannot be called static
'static' should be supported as a class method name, even in strict mode. Currently we eat the static token and just mark a flag.

Added detection similar to async and added a test case. Also included 'let' in the test case, as it is the only other semantically determined keyword.