Remove undefined functions from JavascriptOperators.h
VS2017 marks these functions as not defined. This PR removes the unused function declarations.
The rest three of them are defined in JavascriptOperators.cpp, but the parameter types are not
the same (e.g. int vs int32), so this PR changes the typename so that VS can match the declaration
and the definition.