Fix multi-unit characters after numeric literal
Fixes OS: 14164940

I introduced a regession with my recent change to numeric literal spec complicance. After parsing a numeric literal, we were not checking if there was a multi-byte character before checking it as an IdStart character. In this case, we need to read the full character before checking.

This affected a handful of Test262 tests which used multi-byte trailing whitespace.