Fix int constant folding during inlining for functions with try
We should not int constant fold args during inlining for functions with try.
This bug came up when we enabled inling into functions with try recently.

We maybe const folding across regions, and a BailOnException in the middle will need to restore this.
Int const fold at inline and a later deadstore can result in incorrect results in such a scenario.