Fix clearing callinfo on inlined frames in a try
We inline functions into try, and clear the callinfo of all inlined frames until the enclosing try frame on a throw.
But if the inlined function calls into the interpreter and a throw happens, we do not want to clear the inlinee frame info of outer inlined frames.

In the failing test case, we inlined a function that called into an async function in the interpreter, we incorrectly cleared all the inlined frames until the enclosing try.
And on a following bailout from inside the inlined code in the try, we ended up with cleared callinfo while executing BailOutRecord::BailOutInlinedHelper.

This fix does that stack walk in the catch handler of the JIT TryCatch helper as opposed to doing it in the interpreter. Because right now, we do not mark try frames
of the interpreter on the stack of try-catches we maintain. Doing it in the catch handler we initialize the stack walker from the throwing function and clear
inlinee frames till the most recent enclosing try jit frame.

Fixes OS#14226300