Military time zone calculation is wrong
According to source [1] and [2], the calculation of military
time zone is wrong in ChakraCore. It should be the opposite:
a corresponds to UTC+1, b corresponds to UTC+2, ... e.g.
Eastern standard time is R, or UTC-5.

From [2]:

The suffix indicates the correction (for description, see table) which must be applied to the time as expressed in order to convert to GMT.

Test cases are included.

[1] https://www.timeanddate.com/time/zones/military
[2] https://web.archive.org/web/20160304061301/http://jcs.dtic.mil/j6/cceb/acps/acp121/ACP121I.pdf, pp. 51-52