Clear inlinee callinfo from the catch of OP_TryFinally as well
When we hit an exception and we have a try finally within a try catch. We will execute finally code by calling BailOnException.
Since we inline into functions with try now, we will end up in inconsistent inlinee callinfo when we call the bailout code from finally.
So add the walker and clear inlinee callinfo.

Remove isInlineeCleared bit

I had added isInlineeCleared bit to mark when we have cleared inlinee frames's callinfo count on an exception within a try finally enclosed in some try catch.
I later checked this bit in the enclosing OP_TryCatch to make sure we don't walk the stack to clear info again.

This is not needed, because if there was an expection within try finally, then we will bailout and never exercise the code in OP_TryCatch