treat two digit years less than 50 as 21st century years
One year ago, Mozilla changed the way Firefox interprets dates in the format mm/dd/yy.
In Firefox version 49 or later, if yy is less than 50, the year is regarded as 20yy;
if yy is 50~99, the year is regarded as 19yy. The old Firefox behavior matches IE and Edge,
and its new behavior matches Chrome and Safari.[1]

This PR changes the behavior of parsing two digit years to align with SpiderMonkey and v8.
Tests are included.

[1] https://bugzilla.mozilla.org/show_bug.cgi?id=1265136