Updating the CallInfo count handling logic
Currently for two scenarios (eval and new.target) in call instruction we pass an extra argument (FrameDisplay or New.Target object). This additional count is reflected in the CallInfo.Count field making every callee to be aware of this and update the arg count accordingly. This creates unnecessary confusion at the callee site.

This changelist changes that logic in such a way that callee does not have to be aware of the additional argument in normal scenarios. Only scenarios involving eval and new.target needs to do special handling for the extra arg. This makes CallInfo.Count handling simpler.

For eval calls we currently check whether the module id is the same as the global module id. If not we pass that also as an additional argument. It looks like Edge does not use this value at all and always use the global module id (0). Also the module id field has been repurposed by the ES6 module logic. So I am removing the logic of passing module id from eval call.