Do not go through StartEmitFunction if the fuction is deferred
There can be cases where we are reusing the function body for a child function and we got to StartEmitFunction for the child function when trying to emit the re-deferred parent function. In that case don't redo the slot allocations in StartEmitFunction for the child function as it already has valid bytecode.