Fixing the js reentrancy error happened while throwing.
When we throw we set the no-reentrancy flag to false. However we did that little late. By the time it is possible that we may go out due to stackTraceLimit. Which leads to this Assert.
Fixing this by moving it up. Now this is not real a re-entrancy problem. We were expecting the re-entrancy here due to a throw, it is just that we were resetting the flag little later.