ARM64: ChangeToAssign with Indirs
assigns from IndirOpnds get converted to LDR which can handle 8/16 byte sources. Changing the type of the source can result in the LDR reading some memory off the end of the array, and while those bytes are discarded by the UBFX/SBFX it can result in an access violation if reading at a page boundary.