Relax noisy CheckIsExecutable assert
Fixes OS: 14324283: CAS:EdgeCrawler: ASSERT:  Can't call function when the script context is closed (Chakra!CheckIsExecutable+18b [e:\a\_work\3\s\core\lib\backend\backendapi.cpp @ 139])

F12 is hitting the CheckIsExecutable assert when we have a closed ScriptContext and a call to a CustomExternalObject with a CrossSiteExternalEntryThunk entrypoint.

The fix (without exposing a bunch of ChakraFull things) is to check for a callable external object.