Merged PR 96029: Reduce the scope of a fix to address a perf bug
The change to do this got undone by a fullsync issue when it was done in dev3 (https://microsoft.visualstudio.com/_git/os/commit/6933d4be4f827062c7b62b5f694001efc4891162), and in the meanwhile the macros were removed with the 1711 public release branch merge. In order to get everything into the right state, I'm targeting this at the same branch as the 1711 release.