fix a GC NewTrackBit assertion
In RECYCLER_STATS (debug) mode a newly allocated finalizable heap block has
NewFinalizeBit on (aliased with NewTrackBit and used for finalizable object
counting). This bit is supposed to be cleared later after mark processing.
However these may occur in reversed order. We could falsely mark an address
before the owner heap block allocation, resulting in newly allocated
address marked and NewTrackBit (NewFinalizeBit) on.

This did not repro on Windows because the object usually triggers default
write watch and re-mark of the address. Change the related code to trigger
software write barrier on the address to work around on xplat.

Fix #3280