RL: Apply the timeout to all tests
In the PR https://github.com/Microsoft/ChakraCore/pull/4384 the timeout value was geting applied to only those test cases which has timeout entries in the xml. But from the Nightly runs it looks like tests without a timoeut entry can also timeout. So making this change to apply the timeout to all tests.